Change button design

+++ app/assets/javascripts/practices.js.coffee
@@ -2,30 +2,38 @@ $ ->
   getPracticeId = ($element) ->
     $element.attr('id').split('-')[1]
 
-  $(document).on 'click', '.status.unstarted', ->
-    that = this
-    practice_id = getPracticeId($(this))
+  $('button.unstarted').on 'click', ->
+    return if $(this).hasClass('disabled')
+    practice_id = getPracticeId($(this).parent())
     $.ajax "/practices/#{practice_id}/start",
       type: 'POST'
     .done ->
-      $(that).attr({ class: 'status active' })
-      $(that).children().attr({ class: 'btn btn-info' }).text 'active'
+      $.each $("#practice-#{practice_id}").children(), (i, element)->
+        if $(element).hasClass('active')
+          $(element).removeClass('disabled')
+        else
+          $(element).addClass('disabled')
 
-  $(document).on 'click', '.status.active', ->
-    that = this
-    practice_id = getPracticeId($(this))
+  $('button.active').on 'click', ->
+    return if $(this).hasClass('disabled')
+    practice_id = getPracticeId($(this).parent())
     $.ajax "/practices/#{practice_id}/finish",
       type: 'PUT'
     .done ->
-      $(that).attr({ class: 'status complete' })
-      $(that).children().attr({ class: 'btn btn-success' }).text 'complete'
+      $.each $("#practice-#{practice_id}").children(), (i, element)->
+        if $(element).hasClass('complete')
+          $(element).removeClass('disabled')
+        else
+          $(element).addClass('disabled')
 
-  $(document).on 'click', '.status.complete', ->
-    that = this
-    practice_id = getPracticeId($(this))
+  $('button.complete').on 'click', ->
+    return if $(this).hasClass('disabled')
+    practice_id = getPracticeId($(this).parent())
     $.ajax "/practices/#{practice_id}/destroy",
       type: 'DELETE'
     .done ->
-      $(that).attr({ class: 'status unstarted' })
-      $(that).children().attr({ class: 'btn' }).text 'unstarted'
-
+      $.each $("#practice-#{practice_id}").children(), (i, element)->
+        if $(element).hasClass('unstarted')
+          $(element).removeClass('disabled')
+        else
+          $(element).addClass('disabled')
